requirements.txt
ultralytics
litestar
uvicorn
pillow
